--- docs/expiration_time.mjs
@@ -1,9 +1,0 @@
-export {getMillisecondsToExpiration};
-
-function getMillisecondsToExpiration(expireTime) {
-  return millisecondsFromNow(Date.parse(expireTime));
-}
-
-function millisecondsFromNow(date) {
-  return date - Date.now();
-}
