+++ package.json
@@ -6,12 +6,13 @@
     "@babel/register": "^7.7.7",
     "babel-eslint": "^10.0.3",
     "cypress": "3.6.1",
-    "eslint": "^6.7.1",
+    "eslint": "^6.8.0",
     "eslint-config-google": "^0.14.0",
     "firebase": "^7.6.1",
     "firebase-admin": "^8.9.0",
     "googleapis": "^46.0.0",
     "local-web-server": "^3.0.7"
   },
+  "type": "module",
   "license": "MIT"
 }
